feat(components): update stencil core patch | ks | #3779
