+++ packages/components/package.json
@@ -6,10 +6,9 @@
   "scope": "@porsche-design-system",
   "author": "Dr. Ing. h.c. F. Porsche AG",
   "scripts": {
-    "patchStencilSSRHydration": "tsx scripts/patch-stencil-core/patchStencilSSRHydration.ts",
-    "patchStencilElementInternals": "tsx scripts/patch-stencil-core/patchStencilElementInternals.ts",
-    "patchStencilPropDecorator": "tsx scripts/patch-stencil-core/patchStencilPropDecorator.ts",
-    "patchStencil": "yarn patchStencilSSRHydration && yarn patchStencilElementInternals && yarn patchStencilPropDecorator",
+    "patchStencilCoreCompiler": "tsx scripts/patch-stencil-core/patchStencilCoreCompiler.ts",
+    "patchStencilCoreClient": "tsx scripts/patch-stencil-core/patchStencilCoreClient.ts",
+    "patchStencilCore": "yarn patchStencilCoreClient && yarn patchStencilCoreCompiler",
     "prestart": "yarn patchStencil && yarn prestart:prepareAriaTypes",
     "prestart:prepareAriaTypes": "tsx scripts/prepareAriaTypes.ts",
     "start": "concurrently -k -s 'first' -n 'cdn,stencil' 'serve-cdn' 'PDS_IS_STAGING=1 stencil build --watch --serve --dev'",
